Make TodoItem a separate component
