+++ src/components/index.js
@@ -1,3 +1,4 @@
 export { TodosList } from './TodosList/TodosList';
 export { TodoForm } from './TodoForm/TodoForm';
 export { TodoDetails } from './TodoDetails/TodoDetails';
+export { TodoItem } from './TodoItem/TodoItem'
